requirements.txt
absl-py==2.4.0
alembic==1.18.3
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
astunparse==1.6.3
bcrypt==5.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
cryptography==45.0.7
debugpy==1.8.20
decorator==5.2.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.128.7
flatbuffers==25.12.19
gast==0.4.0
google-auth==2.48.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.3.1
grpcio==1.74.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
ipykernel==7.2.0
ipython==8.38.0
jedi==0.19.2
jupyter_client==8.8.0
jupyter_core==5.9.1
keras==2.13.1
libclang==18.1.1
Mako==1.3.10
Markdown==3.10.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numpy==1.24.3
oauthlib==3.3.1
opt_einsum==3.4.0
packaging==26.0
parso==0.8.5
passlib==1.7.4
platformdirs==4.5.1
pluggy==1.6.0
prompt_toolkit==3.0.52
protobuf==4.25.8
psutil==7.2.2
pure_eval==0.2.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-jose==3.5.0
python-multipart==0.0.22
pyzmq==27.1.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.46
stack-data==0.6.3
starlette==0.52.1
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.3.0
tomli==2.4.0
tornado==6.5.4
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
wcwidth==0.6.0
Werkzeug==3.1.5
wrapt==2.1.1
